Local requirements.txt
# ---------- Core ----------
numpy
scipy

# ---------- Computer Vision ----------
opencv-python-headless
Pillow

# ---------- YOLO ----------
ultralytics

# ---------- Face Recognition (CPU) ----------
tensorflow-cpu
keras-facenet

# ---------- Similarity ----------
scikit-learn

# ---------- AWS ----------
boto3

# ---------- Utilities ----------
tqdm
